+++ examples/data-sources/identitycloud_promotion_lock/data-source.tf
@@ -1,0 +1,2 @@
+data "identitycloud_promotion_lock" "example" {
+}